Add files via upload
In this exercise, you will write a recursive procedure that, given a number N, generates and prints all binary patterns of length N. For example, for N=3, it will output:

000
001
010
011
100
101
110
111
Likewise, if N were 4, the output will be 16 lines long, corresponding to the binary strings of length 4, in increasing order, starting with 0000 and ending with 1111. Assume N will not be more than 20.

This exercise must be done using recursion. C Version & MIPS Assembly Version